# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Родословные",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/login": {
      "post": {
        "summary": "Post Login",
        "operationId": "post_login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/clans/{clan_id}/link": {
      "get": {
        "summary": "Get Share Link",
        "operationId": "get_share_link_api_clans__clan_id__link_get",
        "parameters": [
          {
            "name": "clan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Clan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ShareInfo"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Share Link Api Clans  Clan Id  Link Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Share Link",
        "description": "Выпускает ссылку заново: прежняя перестаёт работать.",
        "operationId": "post_share_link_api_clans__clan_id__link_post",
        "parameters": [
          {
            "name": "clan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Clan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Share Link",
        "description": "Отзывает ссылку: род закрывается и для тех, кто по ней уже приходил.",
        "operationId": "delete_share_link_api_clans__clan_id__link_delete",
        "parameters": [
          {
            "name": "clan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Clan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Share Link Api Clans  Clan Id  Link Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/clans/{clan_id}/tree": {
      "get": {
        "summary": "Get Clan Tree",
        "operationId": "get_clan_tree_api_clans__clan_id__tree_get",
        "parameters": [
          {
            "name": "clan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Clan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClanTree"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/persons/{person_id}": {
      "get": {
        "summary": "Get Person",
        "operationId": "get_person_api_persons__person_id__get",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Person Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Put Person",
        "operationId": "put_person_api_persons__person_id__put",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonFields"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove Person",
        "operationId": "remove_person_api_persons__person_id__delete",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Person Id"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Branch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/persons": {
      "get": {
        "summary": "Find Persons",
        "description": "Поиск по всем родам — для ручной связки.",
        "operationId": "find_persons_api_persons_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "exclude_clan",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exclude Clan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LinkPerson"
                  },
                  "title": "Response Find Persons Api Persons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/clans/{clan_id}/links": {
      "get": {
        "summary": "Get Clan Links",
        "operationId": "get_clan_links_api_clans__clan_id__links_get",
        "parameters": [
          {
            "name": "clan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Clan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClanLink"
                  },
                  "title": "Response Get Clan Links Api Clans  Clan Id  Links Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/links/{link_id}/see": {
      "put": {
        "summary": "Put Link See",
        "description": "Уровень связки: скрытую зритель не видит и перейти по ней не может.",
        "operationId": "put_link_see_api_links__link_id__see_put",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Link Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkSee"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Link"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkSee": {
        "properties": {
          "see": {
            "type": "string",
            "enum": [
              "all",
              "clan",
              "hidden"
            ],
            "title": "See"
          }
        },
        "type": "object",
        "required": [
          "see"
        ],
        "title": "LinkSee"
      },
      "LoginForm": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "name",
          "password"
        ],
        "title": "LoginForm"
      },
...
      "Me": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "guarded": {
            "type": "boolean",
            "title": "Guarded"
          },
          "clans": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Clans",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "guarded"
        ],
        "title": "Me"
      },
...
    }
  }
}